+++ src/main/java/net/inceptioncloud/dragonfly/subscriber/DeveloperModeSubscriber.kt
@@ -18,17 +18,17 @@ object DeveloperModeSubscriber {
         renderDeveloperInfo(
             "FPS: ",
             Minecraft.getDebugFPS().toString(),
-            event.scaledWidth, 2
+            event.scaledWidth, 2.0
         )
         renderDeveloperInfo(
             "TPS: ",
             Dragonfly.lastTPS.toString(),
-            event.scaledWidth, 22
+            event.scaledWidth, 22.0
         )
         renderDeveloperInfo(
             "GUI: ",
             Minecraft.getMinecraft().currentScreen?.javaClass?.simpleName ?: "null",
-            event.scaledWidth, 42
+            event.scaledWidth, 42.0
         )
 
         val asyncBuilding = Dragonfly.fontManager.defaultFont.asyncBuilding
@@ -37,28 +37,28 @@ object DeveloperModeSubscriber {
         renderDeveloperInfo(
             "Building: ",
             asyncBuilding.size.toString(),
-            event.scaledWidth, 62
+            event.scaledWidth, 62.0
         )
         renderDeveloperInfo(
             "Cached: ",
             cachedFontRenderer.size.toString(),
-            event.scaledWidth, 82
+            event.scaledWidth, 82.0
         )
         renderDeveloperInfo(
             "Scaled: ",
             cachedFontRenderer.count { it.value is ScaledFontRenderer }.toString(),
-            event.scaledWidth, 102
+            event.scaledWidth, 102.0
         )
     }
 
-    private fun renderDeveloperInfo(title: String, content: String, screenWidth: Int, y: Int) {
+    private fun renderDeveloperInfo(title: String, content: String, screenWidth: Double, y: Double) {
         val fontRenderer = Dragonfly.fontManager.monospaceFont.fontRenderer(size = 38, useScale = false)
         val height = fontRenderer.height
         val framesTitleWidth = fontRenderer.getStringWidth(title)
         val framesWidth = framesTitleWidth + fontRenderer.getStringWidth(content)
 
         Gui.drawRect(screenWidth - 2 - framesWidth - 2, y, screenWidth - 2, y + height, Color(0, 0, 0, 150).rgb)
-        fontRenderer.drawString(title, screenWidth - 3 - framesWidth, y + 2, Color.WHITE.rgb)
-        fontRenderer.drawString(content, screenWidth - 3 - framesWidth + framesTitleWidth, y + 2, Color.YELLOW.rgb)
+        fontRenderer.drawString(title, (screenWidth - 3 - framesWidth).toInt(), (y + 2).toInt(), Color.WHITE.rgb)
+        fontRenderer.drawString(content, (screenWidth - 3 - framesWidth + framesTitleWidth).toInt(), (y + 2).toInt(), Color.YELLOW.rgb)
     }
 }
